rename string_line_iterator to string::line_iterator
